+++ src/common/utils/poolUtils.ts
@@ -134,7 +134,7 @@ export async function tryFetchScratchOrg(options: any) {
       return null;
     }
     await fs.writeFile(tmpAuthFile, authFileContent, "utf8");
-    const authCommand = `sfdx auth:sfdxurl:store -f ${tmpAuthFile} --setdefaultusername --setalias ${scratchOrg.scratchOrgAlias}`;
+    const authCommand = `sf org login sfdx-url --sfdx-url-file ${tmpAuthFile} --set-default --set-alias ${scratchOrg.scratchOrgAlias}`;
     const authRes = await execSfdxJson(authCommand, this, { fail: false, output: false });
     if (authRes.status !== 0) {
       uxLog(