+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
   - sfdx force:org:delete -> sf org delete scratch
   - sfdx config:get -> sf config get
   - sfdx config:set -> sf config set
+  - sfdx auth:web:login -> sf org login web
+  - sfdx auth:sfdxurl:store -> sf org login sfdx-url
 - Get rid of sfdx-essentials dependency by internalizing its used commands
   - sf hardis:packagexml:append
   - sf hardis:packagexml:remove